Rewriting `TradeAfterThreshold` to work with Safe (#4)
Cf. updated README for explanation and a test plan on why this is desirable.

TradeAfterThreshold was updated to instead of implicitly checking balances for itself it can now check balances of a "target" contract and will sign off orders on behalf of this contract instead of self. This means we can create a new instance for each Safe and use it as a fallbackHandler which will automatically validate trades on behalf of this Safe.

This also means the ConditionalOrder event needs to emit the address of the target contract (as this is the one the watcher needs to monitor)

I built a simple factory contract to allow easy deployment via the Safe UI (cf Readme). I should probably start adding artefact information to the repo (at this point I'm using `forge create` which doesn't record any metadata about the deployment). I will leave this to a later PR.

For now, we can test the deployed factories at:
- https://blockscout.com/xdai/mainnet/address/0x66510332b5EF19086eD5FEB533924ee8fA50f10A
- https://goerli.etherscan.io/address/0x58104dd6844e8aba70b37d69a1afa675f46c40cb